+++ src/app/core/@ngrx/products/products.selectors.ts
@@ -2,7 +2,7 @@ import { createFeatureSelector, createSelector } from '@ngrx/store';
 import { ProductsState } from './products.state';
 import { productsFeatureKey } from './products.reducer';
 import { selectRouterState } from '../router/router.selectors';
-import { ProductModel } from '../../../products';
+import { ProductModel } from '../../../products/models/product.model';
 
 export const selectProductsState = createFeatureSelector<ProductsState>(productsFeatureKey);
 
@@ -19,11 +19,6 @@ export const selectProductsLoaded = createSelector(
 export const selectProductsError = createSelector(
   selectProductsState,
   (state: ProductsState) => state.error
-);
-
-export const selectOriginalProduct = createSelector(
-  selectProductsState,
-  (state: ProductsState) => state.originalProduct
 );
 
 export const selectSelectedProductByUrl = createSelector(