+++ src/app/core/@ngrx/products/products.state.ts
@@ -2,7 +2,6 @@ import { Product } from '../../../products';
 
 export interface ProductsState {
   data: ReadonlyArray<Product>;
-  originalProduct: Readonly<Product>;
   readonly loading: boolean;
   readonly loaded: boolean;
   readonly error: Error | string;
@@ -10,7 +9,6 @@ export interface ProductsState {
 
 export const initialProductsState: ProductsState = {
   data: [],
-  originalProduct: null,
   loading: false,
   loaded: false,
   error: null