+++ src/app/core/@ngrx/router/router.actions.ts
@@ -1,9 +1,0 @@
-import { NavigationExtras } from '@angular/router';
-import { createAction, props } from '@ngrx/store';
-
-export const back = createAction('[Router] BACK');
-
-export const go = createAction(
-  '[Router] GO',
-  props<{ path: any[]; queryParams?: object; extras?: NavigationExtras }>()
-);