+++ src/app/orders/index.ts
@@ -1,5 +1,5 @@
 export * from './components/';
 export * from './model/order.model';
-export * from './services/order.service';
+export * from './services/order-composer.service';