+++ src/app/orders/model/order.model.ts
@@ -1,6 +1,13 @@
 import { CartItem } from '../../cart';
 
-export class OrderModel {
+export interface Order {
+  cartItems: CartItem[];
+  totalPrice: number;
+  totalCount: number;
+  id?: number;
+}
+
+export class OrderModel implements Order {
   constructor(public cartItems: CartItem[],
               public totalPrice: number = 0,
               public totalCount: number = 0,