+++ src/app/orders/orders.module.ts
@@ -2,7 +2,6 @@ import { NgModule } from '@angular/core';
 import { CommonModule } from '@angular/common';
 
 import { OrdersRoutingModule } from './orders-routing.module';
-import { ordersAPIProvider } from './orders.config';
 import { SharedModule } from '../shared/shared.module';
 
 @NgModule({
@@ -11,9 +10,6 @@ import { SharedModule } from '../shared/shared.module';
     CommonModule,
     SharedModule,
     OrdersRoutingModule
-  ],
-  providers: [
-    ordersAPIProvider
   ]
 })
 export class OrdersModule {