+++ src/app/products/components/product-list/product-list.component.ts
@@ -1,13 +1,11 @@
 import { Component, OnInit } from '@angular/core';
 import { Product, ProductModel } from '../../models/product.model';
-import { ProductService } from '../../services/product.service';
 import { CartService } from '../../../shared';
-import { Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { MatSnackBar } from '@angular/material/snack-bar';
 import { MessageSnackbarComponent } from '../../../shared/components';
 import { ProductsFacade } from '../../../core/@ngrx/products/products.facade';
-import { RouterFacade } from '../../../core/@ngrx/router/router.facade';
+import { Router } from '@angular/router';
 
 @Component({
   selector: 'app-product-list',
@@ -20,8 +18,8 @@ export class ProductListComponent implements OnInit {
   error$: Observable<Error | string>;
 
   constructor(private cartService: CartService,
-              private routerFacade: RouterFacade,
               private snackBar: MatSnackBar,
+              private router: Router,
               private productsFacade: ProductsFacade) {
   }
 
@@ -32,6 +30,7 @@ export class ProductListComponent implements OnInit {
 
   onBuy(event: Event, product: ProductModel): void {
     event.preventDefault();
+
     this.cartService.addProduct(product);
     this.snackBar.openFromComponent(MessageSnackbarComponent, {
       data: 'Product was Added to Cart!'
@@ -40,7 +39,8 @@ export class ProductListComponent implements OnInit {
 
   onViewInfo(event: Event, product: ProductModel) {
     event.preventDefault();
+
     const path = ['/products', product.id];
-    this.routerFacade.navigate({ path });
+    this.router.navigate(path);
   }
 }