+++ src/app/products/guards/check-product-existing-guard.service.ts
@@ -1,12 +1,11 @@
 import { Injectable } from '@angular/core';
-import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, UrlTree } from '@angular/router';
+import { CanActivate, ActivatedRouteSnapshot, RouterStateSnapshot, Router } from '@angular/router';
 import { Observable } from 'rxjs';
 import { Store } from '@ngrx/store';
 import { AppState } from '../../core/@ngrx/app.state';
 import { ProductsFacade } from '../../core/@ngrx/products/products.facade';
 import { checkStore } from './check-store.function';
 import { map, switchMap, take, tap } from 'rxjs/operators';
-import { RouterFacade } from '../../core/@ngrx/router/router.facade';
 
 @Injectable({
   providedIn: 'root'
@@ -15,16 +14,13 @@ export class CheckProductExistingGuard implements CanActivate {
 
   constructor(private store: Store<AppState>,
               private productsFacade: ProductsFacade,
-              private routerFacade: RouterFacade) {
+              private router: Router) {
   }
 
-  canActivate(
-    routeSnapshot: ActivatedRouteSnapshot,
-    state: RouterStateSnapshot): Observable<boolean> {
+  canActivate(routeSnapshot: ActivatedRouteSnapshot, state: RouterStateSnapshot): Observable<boolean> {
     return checkStore(this.store).pipe(
       switchMap(() => {
         const id = +routeSnapshot.paramMap.get('productID');
-        console.log(id);
         return this.checkIfTaskExists(id);
       })
     );
@@ -35,7 +31,7 @@ export class CheckProductExistingGuard implements CanActivate {
       map(products => !!products.find(product => product.id === id)),
       tap(result => {
         if (!result) {
-          this.routerFacade.navigate({path: ['/home']});
+          this.router.navigate(['/home']);
         }
       }),
       take(1)