+++ src/app/products/products.module.ts
@@ -1,5 +1,5 @@
 import { NgModule } from '@angular/core';
-import { ProductListComponent } from './components';
+import { ProductListComponent } from './components/product-list/product-list.component';
 import { SharedModule } from '../shared/shared.module';
 import { ProductRoutingModule } from './product-routing.module';
 import { productsApiProvider } from './products.config';